+++ src/hooks.ts
@@ -20,6 +20,9 @@ declare global {
     var context: BrowserContext;
     var page: Page;
     var config: any;
+    var contexts: {
+        [contextName: string]: BrowserContext
+    } | null;
 }
 
 Before(async function () {
@@ -75,8 +78,16 @@ After(async function (scenario: ITestCaseHookParameter) {
         }
     }
     if (global.browser) {
-        await context.close();
-        this.log('browser context closed');
+        if (global.contexts) {
+            for (const contextName in global.contexts) {
+                await global.contexts[contextName].close();
+                this.log(`${contextName} context closed`);
+            }
+            global.contexts = null;
+        } else {
+            await context.close();
+            this.log('browser context closed');
+        }
     }
 });